Correctly parse IntLiteral in MR
When calling a placeholder MR within another MR and using a macro value
(at least, might also affect literal written values), we could not
correctly infer the argument value to pass to the rewritten concrete MR
chosen (as it seems there is something more needed to get the
SourceRange of a macro value). We now search by value and pass that
instead of searching by SourceRanges when dealing with IntegerLiterals.